# apps/compass-ai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Compass AI Backend API",
    "version": "2.1.0",
    "description": "Document analysis with grounded, citation-backed answers, scoped to a client portal. Tenancy is derived from a verified portalFeatureContextToken sent as x-portal-context: NO operation accepts a client, workspace or portal identifier from the caller, and one that carries one is rejected with TENANCY_PROBE. Every operation enforces the derived actor (employee or client) inside its handler. All operations are private: there is no cross-app consumer."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every failure carries a stable machine code. For a portal failure the message is separately authored to be safe to show a client.",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "TENANCY_PROBE",
                  "PORTAL_NOT_BOUND",
                  "FORBIDDEN_CLIENT_ROLE",
                  "PORTAL_VERIFY_FAILED"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "PortalContext": {
        "name": "x-portal-context",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Platform-signed portalFeatureContextToken from the iframe URL. Verified through Gate on every request; a local decode is never trusted. It proves WHERE the request is from, never WHO — identity comes from the caller session."
      },
...
    },
    "responses": {
      "TenancyProbe": {
        "description": "The request carried a client/workspace/portal identifier. Refused rather than ignored.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PortalRejected": {
        "description": "Portal context missing, unverifiable, unbound, paused, or the caller is not a member.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ForbiddenClient": {
        "description": "FORBIDDEN_CLIENT_ROLE — this action is employee-only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/session/settings": {
      "get": {
        "operationId": "getClientSettings",
        "summary": "Read effective settings with provenance",
        "description": "Employee-only. Precedence is portal override, then client settings, then defaults; each key reports which layer it came from.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Settings and month-to-date usage"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      },
      "patch": {
        "operationId": "updateClientSettings",
        "summary": "Update the client settings layer",
        "description": "Employee-only. Writes the client layer rather than the portal override, which would silently shadow it.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated settings"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  }
}